install analyitcs & speed insights
