+++ .eslintrc.json
@@ -56,7 +56,8 @@
         "js": "never",
         "jsx": "never",
         "ts": "never",
-        "tsx": "never"
+        "tsx": "never",
+        "mjs": "never"
       }
     ],
     "import/no-extraneous-dependencies": [
