+++ app/dashboard/layout.tsx
@@ -2,6 +2,48 @@ import createClient from '@/utils/supabase/server';
 import { cookies } from 'next/headers';
 import Link from 'next/link';
 import React from 'react';
+import Tab from './tab';
+
+const LandlordLayout: React.FC<{
+  children: React.ReactNode;
+}> = ({ children }) => {
+  const tabs = [
+    {
+      text: 'Summary',
+      slug: '',
+    },
+    {
+      text: 'Properties',
+      slug: 'properties',
+    },
+    {
+      text: 'Reviews',
+      slug: 'reviews',
+    },
+  ];
+
+  return (
+    <div className='flex w-screen max-w-prose flex-col gap-2 bg-accent/20 rounded-md'>
+      <div className='flex flex-row justify-between items-center p-4 gap-4 bg-foreground/20'>
+        <h1 className='text-xl font-semibold text-accent'>Landlord Dashboard</h1>
+        <div className='flex flex-row gap-4'>
+          <div className='flex flex-wrap items-center gap-2'>
+            {tabs.map((tab) => (
+              <Tab
+                key={`/dashboard${tab.slug}`}
+                item={tab}
+                path='/dashboard'
+              />
+            ))}
+          </div>
+        </div>
+      </div>
+      <div className='p-4'>
+        {children}
+      </div>
+    </div>
+  );
+};
 
 const LandlordRegistrationPropmt = () => (
   <div className='flex flex-col gap-1 items-center'>
@@ -79,7 +121,7 @@ const DashboardLayout = async ({
       </div>
       <main className='flex-1 flex flex-col gap-4'>
         {genericDashboard}
-        {userIsLandlord && landlordDashboard}
+        {userIsLandlord && <LandlordLayout>{landlordDashboard}</LandlordLayout>}
         {userHasLeftReview && tenantDashboard}
         <div className='flex flex-col md:flex-row gap-4'>
           {!userIsLandlord && <LandlordRegistrationPropmt />}
