+++ app/layout.tsx
@@ -1,9 +1,11 @@
-// import { GeistSans } from 'geist/font/sans';
+import { GeistSans } from 'geist/font/sans';
 import './globals.css';
 import NavBar from '@/components/Header/NavBar';
 import Footer from '@/components/Footer/Footer';
 import { NextPage } from 'next';
 import React from 'react';
+import { Analytics } from '@vercel/analytics/react';
+import { SpeedInsights } from '@vercel/speed-insights/next';
 
 const defaultUrl = process.env.VERCEL_URL
   ? `https://${process.env.VERCEL_URL}`
@@ -20,13 +22,16 @@ const Layout: NextPage<{
 }> = ({
   children,
 }) => (
-  <html lang='en'>
+  <html lang='en' className={GeistSans.className}>
     <body className='bg-background text-foreground'>
       <main className='min-h-screen flex flex-col items-center'>
         <NavBar />
         {children}
         <Footer />
       </main>
+
+      <SpeedInsights />
+      <Analytics />
     </body>
   </html>
 );
