+++ package.json
@@ -15,9 +15,12 @@
     "@microsoft/eslint-formatter-sarif": "^2.1.7",
     "@supabase/ssr": "latest",
     "@supabase/supabase-js": "latest",
+    "@vercel/analytics": "^1.2.2",
+    "@vercel/speed-insights": "^1.0.10",
     "autoprefixer": "10.4.15",
     "dotenv": "^16.4.5",
-    "next": "14.1.1-canary.36",
+    "geist": "^1.2.2",
+    "next": "latest",
     "postcss": "8.4.31",
     "react": "18.2.0",
     "react-dom": "18.2.0",
